[INLONG-3068][Manager] Move KafkaOffset to client package
