+++ inlong-manager/manager-client/src/main/java/org/apache/inlong/manager/client/api/KafkaOffset.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.common.enums;
+package org.apache.inlong.manager.client.api;
 
 import lombok.Getter;
